+++ src/GridCal/Engine/Simulations/OPF/ac_opf_ts.py
@@ -18,6 +18,7 @@ This file implements a DC-OPF for time series
 That means that solves the OPF problem for a complete time series at once
 """
 
+from GridCal.Engine.Core.multi_circuit import MultiCircuit
 from GridCal.ThirdParty.pulp import *
 import numpy as np
 from itertools import product
@@ -52,7 +53,7 @@ def add_objective_function(problem: LpProblem,
     problem += f_obj
 
 
-def get_power_injections(C_bus_gen, Pg, C_bus_bat, Pb, C_bus_load, LSlack, Pl):
+def get_power_injections(C_bus_gen, Pg, C_bus_bat, Pb, C_bus_load, PlSlack, QlSlack, Pl, Ql):
     """
     Create the power injections per bus
     :param C_bus_gen: Bus-Generators sparse connectivity matrix (n, ng)
@@ -60,7 +61,8 @@ def get_power_injections(C_bus_gen, Pg, C_bus_bat, Pb, C_bus_load, LSlack, Pl):
     :param C_bus_bat: Bus-Batteries sparse connectivity matrix (n, nb)
     :param Pb: Batteries LpVars (nb, nt)
     :param C_bus_load: Bus-Load sparse connectivity matrix (n, nl)
-    :param LSlack: Load slack LpVars (nl, nt)
+    :param PlSlack: Load (real) slack LpVars (nl, nt)
+    :param QlSlack: Load (imag) slack LpVars (nl, nt)
     :param Pl: Load values (nl, nt)
     :return: Power injection at the buses (n, nt)
     """
@@ -69,17 +71,19 @@ def get_power_injections(C_bus_gen, Pg, C_bus_bat, Pb, C_bus_load, LSlack, Pl):
 
     P += lpDot(C_bus_bat.transpose(), Pb)
 
-    P -= lpDot(C_bus_load.transpose(), Pl - LSlack)
+    P -= lpDot(C_bus_load.transpose(), Pl - PlSlack)
 
-    return P
+    Q = - lpDot(C_bus_load.transpose(), Ql - QlSlack)
+
+    return P, Q
 
 
-def add_nodal_power_balance(numerical_circuit, problem: LpProblem, theta, P, start_, end_):
+def add_nodal_power_balance(numerical_circuit, problem: LpProblem, dvm, dva, P, Q, start_, end_):
     """
     Add the nodal power balance
     :param numerical_circuit: NumericalCircuit instance
     :param problem: LpProblem instance
-    :param theta: Voltage angles LpVars (n, nt)
+    :param dva: Voltage angles LpVars (n, nt)
     :param P: Power injection at the buses LpVars (n, nt)
     :return: Nothing, the restrictions are added to the problem
     """
@@ -92,7 +96,8 @@ def add_nodal_power_balance(numerical_circuit, problem: LpProblem, theta, P, sta
         end_ = len(numerical_circuit.time_array)
     # t = np.arange(start_, end_, 1)
 
-    nodal_restrictions = np.empty((numerical_circuit.nbus, end_ - start_), dtype=object)
+    nodal_restrictions_P = np.empty((numerical_circuit.nbus, end_ - start_), dtype=object)
+    nodal_restrictions_Q = np.empty((numerical_circuit.nbus, end_ - start_), dtype=object)
 
     # for each partition of the profiles...
     for t_key, calc_inputs in calc_inputs_dict.items():
@@ -106,34 +111,60 @@ def add_nodal_power_balance(numerical_circuit, problem: LpProblem, theta, P, sta
 
             # re-pack the variables for the island and time interval
             P_island = P[bus_original_idx, :]  # the sizes already reflect the correct time span
-            theta_island = theta[bus_original_idx, :]  # the sizes already reflect the correct time span
+            Q_island = Q[bus_original_idx, :]  # the sizes already reflect the correct time span
+            dva_island = dva[bus_original_idx, :]  # the sizes already reflect the correct time span
+            dvm_island = dvm[bus_original_idx, :]  # the sizes already reflect the correct time span
             B_island = calculation_input.Ybus[bus_original_idx, :][:, bus_original_idx].imag
+            G_island = calculation_input.Ybus[bus_original_idx, :][:, bus_original_idx].real
+            Bs_island = calculation_input.Yseries[bus_original_idx, :][:, bus_original_idx].imag
+            Gs_island = calculation_input.Yseries[bus_original_idx, :][:, bus_original_idx].real
 
             pqpv = calculation_input.pqpv
+            pq = calculation_input.pq
+            pv = calculation_input.pv
             vd = calculation_input.ref
+            vdpv = np.r_[vd, pv]
+            vdpv.sort()
 
-            # Add nodal power balance for the non slack nodes
-            nodal_restrictions[pqpv] = lpAddRestrictions2(problem=problem,
-                                                          lhs=lpDot(B_island[pqpv, :][:, pqpv], theta_island[pqpv, :]),
-                                                          rhs=P_island[pqpv, :],
-                                                          name='Nodal_power_balance_pqpv',
+            # Add nodal real power balance for the non slack nodes
+            nodal_restrictions_P[pqpv] = lpAddRestrictions2(problem=problem,
+                                                            lhs=-lpDot(Bs_island[pqpv, :][:, pqpv], dva_island[pqpv, :])
+                                                                + lpDot(G_island[pqpv, :][:, pq], dvm_island[pq, :]),
+                                                            rhs=P_island[pqpv, :],
+                                                            name='Nodal_real_power_balance_pqpv',
+                                                            op='=')
+
+            # Add nodal reactive power balance for the non slack nodes
+            nodal_restrictions_Q[pq] = lpAddRestrictions2(problem=problem,
+                                                          lhs=-lpDot(Gs_island[pq, :][:, pqpv], dva_island[pqpv, :])
+                                                              - lpDot(B_island[pq, :][:, pq], dvm_island[pq, :]),
+                                                          rhs=Q_island[pq, :],
+                                                          name='Nodal_imag_power_balance_pqpv',
                                                           op='=')
 
-            # Add nodal power balance for the slack nodes
-            nodal_restrictions[vd] = lpAddRestrictions2(problem=problem,
-                                                        lhs=lpDot(B_island[vd, :], theta_island),
-                                                        rhs=P_island[vd, :],
-                                                        name='Nodal_power_balance_vd',
-                                                        op='=')
+            # Add nodal real power balance for the slack nodes
+            nodal_restrictions_P[pqpv] = lpAddRestrictions2(problem=problem,
+                                                            lhs=-lpDot(Bs_island[vd, :], dva_island)
+                                                                + lpDot(G_island[vd, :], dvm_island),
+                                                            rhs=P_island[vd, :],
+                                                            name='Nodal_real_power_balance_vd',
+                                                            op='=')
 
-            # slack angles equal to zero
+            # delta of voltage angles equal to zero for the slack nodes (vd)
             lpAddRestrictions2(problem=problem,
-                               lhs=theta_island[vd, :],
-                               rhs=np.zeros_like(theta_island[vd, :]),
-                               name='Theta_vd_zero',
+                               lhs=dva_island[vd, :],
+                               rhs=np.zeros_like(dva_island[vd, :]),
+                               name='dVa_vd_zero',
                                op='=')
 
-    return nodal_restrictions
+            # delta of voltage module equal to zero for the slack and pv nodes (vdpv)
+            lpAddRestrictions2(problem=problem,
+                               lhs=dvm_island[vdpv, :],
+                               rhs=np.zeros_like(dva_island[vdpv, :]),
+                               name='dVm_vdpv_zero',
+                               op='=')
+
+    return nodal_restrictions_P, nodal_restrictions_Q
 
 
 def add_branch_loading_restriction(problem: LpProblem,
@@ -209,9 +240,9 @@ def add_battery_discharge_restriction(problem: LpProblem, SoC0, Capacity, Effici
                            op='=')
 
 
-class OpfDcNonSequentialTimeSeries:
+class OpfAcNonSequentialTimeSeries:
 
-    def __init__(self, grid, start_idx, end_idx):
+    def __init__(self, grid: MultiCircuit, start_idx, end_idx):
         """
 
         :param grid:
@@ -222,10 +253,13 @@ class OpfDcNonSequentialTimeSeries:
         self.start_idx = start_idx
         self.end_idx = end_idx
 
-        self.theta = None
+        self.v0 = None
+        self.dva = None
+        self.dvm = None
         self.Pg = None
         self.Pb = None
         self.Pl = None
+        self.Ql = None
         self.E = None
         self.s_from = None
         self.s_to = None
@@ -254,6 +288,9 @@ class OpfDcNonSequentialTimeSeries:
         b = self.end_idx
         Sbase = numerical_circuit.Sbase
 
+        # bus
+        self.v0 = np.abs(numerical_circuit.V0)
+
         # battery
         Capacity = numerical_circuit.battery_Enom / Sbase
         minSoC = numerical_circuit.battery_min_soc
@@ -271,12 +308,12 @@ class OpfDcNonSequentialTimeSeries:
 
         # load
         Pl = (numerical_circuit.load_active_prof[a:b, :] * numerical_circuit.load_power_profile.real[a:b, :]).transpose() / Sbase
+        Ql = (numerical_circuit.load_active_prof[a:b, :] * numerical_circuit.load_power_profile.imag[a:b, :]).transpose() / Sbase
         cost_l = numerical_circuit.load_cost_prof[a:b, :].transpose()
 
         # branch
         branch_ratings = numerical_circuit.br_rates / Sbase
-        Bseries = (numerical_circuit.branch_active_prof[a:b, :] * (
-                    1 / (numerical_circuit.R + 1j * numerical_circuit.X))).imag.transpose()
+        Bseries = (numerical_circuit.branch_active_prof[a:b, :] * (1 / (numerical_circuit.R + 1j * numerical_circuit.X))).imag.transpose()
         cost_br = numerical_circuit.branch_cost_profile[a:b, :].transpose()
 
         # Compute time delta in hours
@@ -289,9 +326,10 @@ class OpfDcNonSequentialTimeSeries:
         Pb = lpMakeVars(name='Pb', shape=(nb, nt), lower=Pb_min, upper=Pb_max)
         E = lpMakeVars(name='E', shape=(nb, nt), lower=Capacity * minSoC, upper=Capacity * maxSoC)
         load_slack = lpMakeVars(name='LSlack', shape=(nl, nt), lower=0, upper=None)
-        theta = lpMakeVars(name='theta', shape=(n, nt), lower=-3.14, upper=3.14)
-        theta_f = theta[numerical_circuit.F, :]
-        theta_t = theta[numerical_circuit.T, :]
+        dva = lpMakeVars(name='dva', shape=(n, nt), lower=-3.14, upper=3.14)
+        dvm = lpMakeVars(name='dvm', shape=(n, nt), lower=0, upper=2)
+        theta_f = dva[numerical_circuit.F, :]
+        theta_t = dva[numerical_circuit.T, :]
         branch_rating_slack1 = lpMakeVars(name='FSlack1', shape=(m, nt), lower=0, upper=None)
         branch_rating_slack2 = lpMakeVars(name='FSlack2', shape=(m, nt), lower=0, upper=None)
 
@@ -302,16 +340,18 @@ class OpfDcNonSequentialTimeSeries:
         add_objective_function(problem, Pg, Pb, load_slack, branch_rating_slack1, branch_rating_slack2,
                                cost_g, cost_b, cost_l, cost_br)
 
-        P = get_power_injections(C_bus_gen=numerical_circuit.C_gen_bus,
-                                 Pg=Pg,
-                                 C_bus_bat=numerical_circuit.C_batt_bus,
-                                 Pb=Pb,
-                                 C_bus_load=numerical_circuit.C_load_bus,
-                                 LSlack=load_slack,
-                                 Pl=Pl)
+        # compute the power injections per node
+        P, Q = get_power_injections(C_bus_gen=numerical_circuit.C_gen_bus, Pg=Pg,
+                                    C_bus_bat=numerical_circuit.C_batt_bus, Pb=Pb,
+                                    C_bus_load=numerical_circuit.C_load_bus,
+                                    PlSlack=load_slack, QlSlack=load_slack,
+                                    Pl=Pl, Ql=Ql)
 
-        nodal_restrictions = add_nodal_power_balance(numerical_circuit, problem, theta, P,
-                                                     start_=self.start_idx, end_=self.end_idx)
+        # compute the nodal power balance restrictions
+        nodal_restrictions_P, nodal_restrictions_Q = add_nodal_power_balance(numerical_circuit=numerical_circuit,
+                                                                             problem=problem,
+                                                                             dvm=dvm, dva=dva, P=P, Q=Q,
+                                                                             start_=self.start_idx, end_=self.end_idx)
 
         load_f, load_t = add_branch_loading_restriction(problem, theta_f, theta_t, Bseries, branch_ratings,
                                                         branch_rating_slack1, branch_rating_slack2)
@@ -322,7 +362,8 @@ class OpfDcNonSequentialTimeSeries:
 
         # Assign variables to keep
         # transpose them to be in the format of GridCal: time, device
-        self.theta = theta.transpose()
+        self.dva = dva.transpose()
+        self.dvm = dvm.transpose()
         self.Pg = Pg.transpose()
         self.Pb = Pb.transpose()
         self.Pl = Pl.transpose()
@@ -332,7 +373,7 @@ class OpfDcNonSequentialTimeSeries:
         self.s_to = load_t.transpose()
         self.overloads = (branch_rating_slack1 + branch_rating_slack2).transpose()
         self.rating = branch_ratings
-        self.nodal_restrictions = nodal_restrictions
+        self.nodal_restrictions = nodal_restrictions_P
 
         return problem
 
@@ -340,6 +381,8 @@ class OpfDcNonSequentialTimeSeries:
         """
         Call PuLP to solve the problem
         """
+        # self.problem.writeLP('AC_OPF.lp')
+
         params = PULP_CBC_CMD(fracGap=0.00001, threads=None, msg=1)
         self.problem.solve(params)
 
@@ -365,8 +408,9 @@ class OpfDcNonSequentialTimeSeries:
         return the complex voltages (time, device)
         :return: 2D array
         """
-        angles = self.extract2D(self.theta)
-        return np.ones_like(angles) * np.exp(-1j * angles)
+        angles = self.extract2D(self.dva)
+        modules = self.v0 + self.extract2D(self.dvm)
+        return modules * np.exp(-1j * angles)
 
     def get_overloads(self):
         """
@@ -434,7 +478,8 @@ class OpfDcNonSequentialTimeSeries:
         """
         val = np.zeros(self.nodal_restrictions.shape)
         for i, j in product(range(val.shape[0]), range(val.shape[1])):
-            val[i, j] = - self.nodal_restrictions[i, j].pi
+            if self.nodal_restrictions[i, j] is not None:
+                val[i, j] = - self.nodal_restrictions[i, j].pi
         return val.transpose()
 
 
@@ -443,12 +488,13 @@ if __name__ == '__main__':
         from GridCal.Engine.IO.file_handler import FileOpen
 
         # fname = '/home/santi/Documentos/GitHub/GridCal/Grids_and_profiles/grids/Lynn 5 Bus pv.gridcal'
+        fname = '/home/santi/Documentos/GitHub/GridCal/Grids_and_profiles/grids/IEEE39_1W.gridcal'
         # fname = r'C:\Users\A487516\Documents\GitHub\GridCal\Grids_and_profiles\grids\Lynn 5 Bus pv.gridcal'
         # fname = r'C:\Users\A487516\Documents\GitHub\GridCal\Grids_and_profiles\grids\IEEE39_1W.gridcal'
-        fname = '/home/santi/Documentos/GitHub/GridCal/Grids_and_profiles/grids/IEEE39_1W.gridcal'
+
         main_circuit = FileOpen(fname).open()
 
-        problem = OpfAcNonSequentialTimeSeries(grid=main_circuit, start_idx=5, end_idx=5 + 5*24)
+        problem = OpfAcNonSequentialTimeSeries(grid=main_circuit, start_idx=1, end_idx=24*5)
 
         print('Solving...')
         status = problem.solve()
@@ -456,6 +502,7 @@ if __name__ == '__main__':
         # print("Status:", status)
 
         v = problem.get_voltage()
+        print('Modules\n', np.abs(v))
         print('Angles\n', np.angle(v))
 
         l = problem.get_loading()
@@ -463,5 +510,8 @@ if __name__ == '__main__':
 
         g = problem.get_generator_power()
         print('Gen power\n', g)
+
+        pr = problem.get_shadow_prices()
+        print('Nodal prices \n', pr)
 
         pass